use one ExpectEventsMismatch exception
